Fixed addPermission and removePermission methods

+++ src/bbn/User/Manager.php
@@ -1090,13 +1090,22 @@ You can click the following link to access directly your account:<br>
       throw new \Exception("No paraneters!");
     }
 
+    if (!($pref = \bbn\User\Preferences::getInstance())) {
+      throw new \Exception("No User\Preferences instance!");
+    }
+
+    if (!($prefCfg = $pref->getClassCfg())) {
+      throw new \Exception("No User\Preferences cfg!");
+    }
+
+
     return (bool)$this->db->insertIgnore(
-      'bbn_users_options',
+      $prefCfg['tables']['user_options'],
       [
-        'id_option' => $id_perm,
-        'id_user' => $id_user,
-        'id_group' => $id_group,
-        'public' => $public
+        $prefCfg['arch']['user_options']['id_option'] => $id_perm,
+        $prefCfg['arch']['user_options']['id_user'] => $id_user,
+        $prefCfg['arch']['user_options']['id_group'] => $id_group,
+        $prefCfg['arch']['user_options']['public'] => $public
       ]
     );
   }
@@ -1108,13 +1117,21 @@ You can click the following link to access directly your account:<br>
       throw new \Exception("No paraneters!");
     }
 
+    if (!($pref = \bbn\User\Preferences::getInstance())) {
+      throw new \Exception("No User\Preferences instance!");
+    }
+
+    if (!($prefCfg = $pref->getClassCfg())) {
+      throw new \Exception("No User\Preferences cfg!");
+    }
+
     return (bool)$this->db->deleteIgnore(
-      'bbn_users_options',
+      $prefCfg['tables']['user_options'],
       [
-        'id_option' => $id_perm,
-        'id_user' => $id_user,
-        'id_group' => $id_group,
-        'public' => $public
+        $prefCfg['arch']['user_options']['id_option'] => $id_perm,
+        $prefCfg['arch']['user_options']['id_user'] => $id_user,
+        $prefCfg['arch']['user_options']['id_group'] => $id_group,
+        $prefCfg['arch']['user_options']['public'] => $public
       ]
     );
   }
